Fixes issue with space in repo path. Fixes #2804
